requirements.txt
aiobotocore==2.11.2
aiohttp==3.9.0
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
botocore==1.34.34
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
dataclasses-json==0.6.2
decorator==5.1.1
dill==0.3.8
distlib==0.3.7
distro==1.8.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.0
fsspec==2023.10.0
google-api-core==2.17.1
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.20.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.339
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langchain-pinecone==0.0.2
langsmith==0.0.87
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.2.2
openai==1.12.0
orjson==3.9.10
packaging==23.2
pandas==2.2.0
pinecone-client==3.0.3
platformdirs==4.1.0
protobuf==4.25.3
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3fs==2023.12.2
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
tqdm==4.66.1
transformers==4.37.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.3
